chore: renamed protocol to identify, removed comments
