+++ setup.py
@@ -117,6 +117,7 @@ setup(
             "chat-demo=examples.chat.chat:main",
             "echo-demo=examples.echo.echo:main",
             "ping-demo=examples.ping.ping:main",
+            "identify-demo=examples.identify.identify:main",
         ],
     },
 )
